move transformLink to common
